interface refactor + context for defaults
