+++ packages/ast/src/encoding/proto/index.ts
@@ -5,3 +5,4 @@ export * from './types';
 export * from './from-json';
 export * from './from-partial';
 export * from './to-json';
+export * from './interface';
