+++ packages/ast/types/src/encoding/proto/index.d.ts
@@ -5,3 +5,4 @@ export * from './types';
 export * from './from-json';
 export * from './from-partial';
 export * from './to-json';
+export * from './interface';
