+++ packages/ast/types/src/encoding/proto/interface.d.ts
@@ -1,0 +1,5 @@
+import * as t from '@babel/types';
+import { ProtoType } from '@osmonauts/types';
+import { ProtoParseContext } from '../context';
+export declare const createProtoType: (context: ProtoParseContext, name: string, proto: ProtoType) => t.ExportNamedDeclaration;
+export declare const createCreateProtoType: (context: ProtoParseContext, name: string, proto: ProtoType) => t.FunctionDeclaration;
