+++ packages/ast/types/src/encoding/proto/types.d.ts
@@ -80,7 +80,7 @@ export declare const types: {
 export declare const getWireNumber: (type: any) => any;
 export declare const getPackedWireNumber: (type: any) => any;
 export declare const getTagNumber: (field: ProtoField) => number;
-export declare const getDefaultTSTypeFromProtoType: (field: ProtoField, isOptional: boolean) => t.ArrayExpression | t.BooleanLiteral | t.Identifier | t.MemberExpression | t.NewExpression | t.NumericLiteral | t.ObjectExpression | t.StringLiteral;
+export declare const getDefaultTSTypeFromProtoType: (context: ProtoParseContext, field: ProtoField, isOptional: boolean) => t.ArrayExpression | t.BooleanLiteral | t.Identifier | t.MemberExpression | t.NewExpression | t.NumericLiteral | t.ObjectExpression | t.StringLiteral;
 export declare const getEnumToJsonName: (name: any) => string;
 export declare const getEnumFromJsonName: (name: any) => string;
 export declare const getFieldsTypeName: (field: ProtoField) => string;
@@ -89,5 +89,3 @@ export declare const getBaseCreateTypeFuncName: (name: any) => string;
 export declare const getOneOfs: (type: ProtoType) => string[];
 export declare const getFieldOptionality: (field: ProtoField, isOneOf: boolean) => boolean;
 export declare const getObjectNameOld: (name: string, scope?: string[]) => string;
-export declare const createProtoType: (context: ProtoParseContext, name: string, proto: ProtoType) => t.ExportNamedDeclaration;
-export declare const createCreateProtoType: (name: string, proto: ProtoType) => t.FunctionDeclaration;
